+++ src/main/java/com/pf/healthybox/dto/orderInformationDto/OiSubscribeBasketDto.java
@@ -12,12 +12,11 @@ public record OiSubscribeBasketDto(
         String sellerCode,
         int productIdx,
         int qty,
-        LocalDateTime deliveryDate,
-        LocalDateTime createdAt
+        LocalDateTime deliveryDate
 ) {
 
-    public static OiSubscribeBasketDto of(String basketNo, String userId, String subscribeCode, String productCode, String sellerCode, int productIdx, int qty, LocalDateTime deliveryDate, LocalDateTime createdAt) {
-        return new OiSubscribeBasketDto(basketNo, userId, subscribeCode, productCode, sellerCode, productIdx, qty, deliveryDate, createdAt);
+    public static OiSubscribeBasketDto of(String basketNo, String userId, String subscribeCode, String productCode, String sellerCode, int productIdx, int qty, LocalDateTime deliveryDate) {
+        return new OiSubscribeBasketDto(basketNo, userId, subscribeCode, productCode, sellerCode, productIdx, qty, deliveryDate);
     }
 
     // dto >> entity
@@ -27,7 +26,7 @@ public record OiSubscribeBasketDto(
 
     // entity >> dto
     public static OiSubscribeBasketDto from(OiSubscribeBasket entity) {
-        return OiSubscribeBasketDto.of(entity.getBasketNo(), entity.getUserId(), entity.getSubscribeCode(), entity.getProductCode(), entity.getSellerCode(), entity.getProductIdx(), entity.getQty(), entity.getDeliveryDate(), entity.getCreatedAt());
+        return OiSubscribeBasketDto.of(entity.getBasketNo(), entity.getUserId(), entity.getSubscribeCode(), entity.getProductCode(), entity.getSellerCode(), entity.getProductIdx(), entity.getQty(), entity.getDeliveryDate());
     }
 
 }
